fix: исправления и изменения в ТЗ-10

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -7,7 +7,6 @@ import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -44,37 +43,55 @@ public class UserService {
     }
 
     public void deleteUser(int id) {
-        for (Long friendsId : userStorage.findUser(id).getFriends()) {
-            userStorage.findUser(friendsId).getFriends().removeIf(tempId -> tempId == id);
+        User userToDelete = userStorage.findUser(id);
+        if (userToDelete != null) {
+            for (Long friendId : userToDelete.getFriends()) {
+                User friend = userStorage.findUser(friendId);
+                if (friend != null) {
+                    friend.getFriends().removeIf(tempId -> tempId == id);
+                }
+            }
+            userStorage.deleteUser(id);
         }
-        userStorage.deleteUser(id);
     }
 
     public void addFriend(int id, int friendId) {
-        userStorage.findUser(id).getFriends().add((long) friendId);
-        userStorage.findUser(friendId).getFriends().add((long) id);
-        log.info("Пользователи с id: {} и {} стали друзьями.", id, friendId);
+        User user = userStorage.findUser(id);
+        User friend = userStorage.findUser(friendId);
+
+        if (user != null && friend != null) {
+            user.getFriends().add((long) friendId);
+            friend.getFriends().add((long) id);
+            log.info("Пользователи с id: {} и {} стали друзьями.", id, friendId);
+        }
     }
 
     public void deleteFriend(int id, int friendId) {
-        userStorage.findUser(id).getFriends().remove(userStorage.findUser(friendId).getId());
-        userStorage.findUser(friendId).getFriends().remove(userStorage.findUser(id).getId());
-        log.info("Пользователь удален из списка друзей.");
+        User user = userStorage.findUser(id);
+        User friend = userStorage.findUser(friendId);
+
+        if (user != null && friend != null) {
+            user.getFriends().remove(user.getId());
+            friend.getFriends().remove(user.getId());
+            log.info("Пользователь удален из списка друзей.");
+        }
     }
 
     public List<User> getFriends(int id) {
-        List<User> userFriends = new ArrayList<>();
-        findAllUsers().forEach(user -> userStorage.findUser(id).getFriends().stream()
-                .filter(idList -> user.getId() == idList)
-                .map(idList -> user)
-                .forEach(userFriends::add));
-        log.info("Количество друзей у пользователя {}: {}.", userStorage.findUser(id).getName(), userFriends.size());
+        User user = userStorage.findUser(id);
+        List<User> userFriends = findAllUsers().stream()
+                .filter(u -> user.getFriends().contains(u.getId()))
+                .collect(Collectors.toList());
+
+        log.info("Количество друзей у пользователя {}: {}.", user.getName(), userFriends.size());
         return userFriends;
     }
 
     public List<User> getCommonFriends(int id, int otherId) {
-        List<User> commonFriends = getFriends(id).stream()
-                .filter(getFriends(otherId)::contains)
+        List<User> userFriends = getFriends(id);
+        List<User> otherUserFriends = getFriends(otherId);
+        List<User> commonFriends = userFriends.stream()
+                .filter(otherUserFriends::contains)
                 .collect(Collectors.toList());
         log.info("Количество общих друзей: {}.", commonFriends.size());
         return commonFriends;
@@ -90,5 +107,4 @@ public class UserService {
             log.info("Автоматически добавлено имя пользователя: {}.", user.getName());
         }
     }
-
 }
